Move C++ helper functions to their own source files.
Create cerr_printf.cc and read_to_list.cc.

Alter main.cc and Makefile.am for the new files.

I considered adding these functions to libdispatch.a and experimented
with that configuration.  I decided that it was not a good idea to mix
C and C++ code in a single library, even if it works.

This change is not worthy of a release tag, IMHO.

Signed-off-by: Jason Stephenson <jason@sigio.com>