Add package URL and add a to version string.
This branch is for a bug fix release that does not largely affect
the actual code. It merely updates the man page and some package
metadata.

If anyone does any further development of the 1.0 branch, it should
be continued from here. As for me, I am already working on a 2.0
branch and release. It will be ready when I have added gettext and
libboost successfully.

Signed-off-by: Jason Stephenson <jason@sigio.com>