Begin work for the 2.0.0 release/branch of dispatch.
Use GNU libtool to build a shared library for the dispatch function.
Make the necessary changes to the source code so that the dispatch
function will be used from the shared library.

Add a man page for the dispatch() function in libdispatch.

Alter the section 1 man page for the dispatch utility.

Update the Makefile.am files for the changes.

Make the appropriate changes to configure.ac to invoke libtool.
Also alter configure.ac for some recent changes to best practices
by add AC_CONFIG_AUX_DIR and AC_CONFIG_MACRO_DIR macro calls.

Signed-off-by: Jason Stephenson <jason@sigio.com>