Add list_sink to read_to_list.cc.
Implement a specific subclass of boost::iostreams::sink to write to
a std::list<char *>.  This simplifies the logic of read_to_list()
somewhat and suggest a name change of copy_to_list, but I won't do
that right now.  (I have considered renaming the function to
read_command(), since that is its ultimate purpose.)

The new class is not as generic as one might like, since it only
works with one specialization of a specific template class.  However,
it is small (about a dozen lines) and fits our needs to a T.

Signed-off-by: Jason Stephenson <jason@sigio.com>