Add a command input filter using boost::iostreams::input_filter.
This input filter allows more complexity in the command file than
just a simple list of lines.  Comments are now permitted and \ can
be used to indicate line continuation.  Whitespace is effectively
skipped.

Update po/dispatch.pot for changed line numbers in source files.

This represents the final step of the boostification effort.

Signed-off-by: Jason Stephenson <jason@sigio.com>