Update files for release 2.0.1.
The dispatch function in libdispatch was corrected to use the
current environment, so this warrants a bug fix release.

The README, configure.ac, lib/Makefile.am, and po/dispatch.pot
files were all updated to reflect the version of the new release
and to document the change in the library function's behavior.

Signed-off-by: Jason Stephenson <jason@sigio.com>